requirements.txt
streamlit==1.28.1
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.32
PyPDF2==3.0.1
tiktoken==0.5.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
chromadb==0.4.22
openai>=1.6.1,<2.0.0
langchain-chroma==0.1.0
thefuzz
SpeechRecognition
pyaudio
streamlit-mic-recorder
streamlit-js-eval